Move HttpUrlConnection helpers to bootstrap
