+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/instrumentation/httpurlconnection/HeadersInjectAdapter.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.http_url_connection;
+package datadog.trace.bootstrap.instrumentation.httpurlconnection;
 
 import datadog.trace.bootstrap.instrumentation.api.AgentPropagation;
 import java.net.HttpURLConnection;
